+++ src/app/admin/user-profile/user-profile.component.ts
@@ -4,7 +4,10 @@ import { AngularFireAuth } from '@angular/fire/auth';
 import { Observable } from 'rxjs';
 
 import { UserService } from '@services/users.service';
+import { AuditService } from '@services/audit.service';
 import { IUser } from '@models/user';
+import { IAuditItem } from '@models/audit';
+import { BaseType } from '@models/base';
 
 @Component({
   selector: 'app-user-profile',
@@ -14,9 +17,13 @@ import { IUser } from '@models/user';
 export class UserProfileComponent implements OnInit {
 
   public userData$: Observable<IUser>;
+  public auditItems: IAuditItem[] = [];
+  public baseTypeEntity: BaseType = BaseType.ENTITY;
+  public baseTypeAudit: BaseType = BaseType.AUDIT;
 
   constructor(
     public auth: AngularFireAuth,
+    private auditSrv: AuditService,
     private userSrv: UserService
   ) { }
 
@@ -24,6 +31,14 @@ export class UserProfileComponent implements OnInit {
     this.auth.user.subscribe((usr) => {
       const uidUser = usr.uid;
       this.userData$ = this.userSrv.getOneUser(uidUser);
+      this.getAudit(usr.uid);
+    });
+  }
+
+  getAudit(uidUser: string): void {
+    this.auditSrv.getAllAuditItemsByUser(uidUser)
+    .subscribe( (auditItems: IAuditItem[]) => {
+      this.auditItems = auditItems;
     });
   }
 }